# Performance improvment: Add format "raw" for "string" type (skip escaping)