--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -113,7 +113,9 @@ module.exports = class Serializer {
       }
     }
 
-    if (str.length < 42) {
+    if (format === 'raw') {
+      return '"' + str + '"';
+    } else if (str.length < 42) {
       return this.asStringSmall(str)
     } else if (STR_ESCAPE.test(str) === false) {
       return '"' + str + '"'