--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -114,7 +114,7 @@ module.exports = class Serializer {
     }
 
     if (format === 'raw') {
-      return '"' + str + '"';
+      return '"' + str + '"'
     } else if (str.length < 42) {
       return this.asStringSmall(str)
     } else if (STR_ESCAPE.test(str) === false) {