--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -98,7 +98,7 @@ module.exports = class Serializer {
     throw new Error(`The value "${date}" cannot be converted to a time.`)
   }
 
-  asString (str) {
+  asString (str, format) {
     if (typeof str !== 'string') {
       if (str === null) {
         return '""'