JSON with a lot of strings take time during escaping. 

This PR introduce "raw" string format (no need to escape).

```json
"schema": {
  "type": "string",
  "format": "raw"
}
```
benchmark (compare classic bench VS same bench with format raw)
```
short string............................................. x 22,581,464 ops/sec ±0.79% (186 runs sampled)
short string and raw..................................... x 1,038,659,823 ops/sec ±0.40% (189 runs sampled)

long string without double quotes........................ x 22,477 ops/sec ±0.46% (192 runs sampled)
long string without double quotes raw.................... x 1,035,763,724 ops/sec ±0.39% (189 runs sampled)

long string.............................................. x 15,865 ops/sec ±0.42% (190 runs sampled)
long string raw.......................................... x 1,060,214,143 ops/sec ±0.49% (190 runs sampled)
```

raw string seem x45 faster