@cesco69 2024-03-08
> Can you add some docs?

Done. https://github.com/fastify/fast-json-stringify/pull/685/commits/4f30811a7eb947122c592776c997b3ae2cd8aa11